+++ apps/meteor/client/views/room/contextualBar/uikit/UiKitContextualBar.tsx
@@ -1,4 +1,3 @@
-import type { UiKit } from '@rocket.chat/core-typings';
 import { Avatar, Box, Button, ButtonGroup, ContextualbarFooter, ContextualbarHeader, ContextualbarTitle } from '@rocket.chat/fuselage';
 import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
 import {
@@ -7,6 +6,7 @@ import {
 	contextualBarParser,
 	UiKitContext,
 } from '@rocket.chat/fuselage-ui-kit';
+import type * as UiKit from '@rocket.chat/ui-kit';
 import { BlockContext } from '@rocket.chat/ui-kit';
 import type { FormEvent, UIEvent } from 'react';
 import React, { memo } from 'react';