+++ apps/meteor/server/services/startup.ts
@@ -16,6 +16,7 @@ import { LDAPService } from './ldap/service';
 import { MessageService } from './messages/service';
 import { MeteorService } from './meteor/service';
 import { NPSService } from './nps/service';
+import { OmnichannelAnalyticsService } from './omnichannel-analytics/service';
 import { OmnichannelIntegrationService } from './omnichannel-integrations/service';
 import { OmnichannelVoipService } from './omnichannel-voip/service';
 import { OmnichannelService } from './omnichannel/service';
@@ -57,6 +58,7 @@ api.registerService(new TranslationService());
 api.registerService(new SettingsService());
 api.registerService(new OmnichannelIntegrationService());
 api.registerService(new ImportService());
+api.registerService(new OmnichannelAnalyticsService());
 
 // if the process is running in micro services mode we don't need to register services that will run separately
 if (!isRunningMs()) {