+++ apps/meteor/tests/end-to-end/api/livechat/04-dashboards.ts
@@ -784,7 +784,7 @@ describe('LIVECHAT - dashboards', function () {
 				// { title: 'Total_messages', value: 6 },
 				// { title: 'Busiest_day', value: moment().format('dddd') }, // TODO: need to check y this return a day before
 				{ title: 'Conversations_per_day', value: '3.50' },
-				{ title: 'Busiest_time', value: '- -' },
+				{ title: 'Busiest_time', value: '' },
 			];
 
 			expectedResult.forEach((expected) => {